+++ src/main/java/com/zero/triptalk/application/PlannerApplication.java
@@ -11,7 +11,9 @@ import com.zero.triptalk.like.service.LikeService;
 import com.zero.triptalk.place.entity.Place;
 import com.zero.triptalk.place.entity.PlaceRequest;
 import com.zero.triptalk.place.service.PlaceService;
-import com.zero.triptalk.planner.dto.*;
+import com.zero.triptalk.planner.dto.request.*;
+import com.zero.triptalk.planner.dto.response.PlannerDetailResponse;
+import com.zero.triptalk.planner.dto.response.PlannerResponse;
 import com.zero.triptalk.planner.entity.Planner;
 import com.zero.triptalk.planner.entity.PlannerDetail;
 import com.zero.triptalk.planner.service.PlannerDetailService;
