--- src/main/java/com/zero/triptalk/planner/dto/CreatePlannerInfo.java
@@ -1,18 +1,0 @@
-package com.zero.triptalk.planner.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class CreatePlannerInfo {
-
-    private List<PlannerDetailListRequest> plannerDetailListRequests;
-    private PlannerRequest plannerRequest;
-}
