+++ src/main/java/com/zero/triptalk/planner/dto/request/PlannerDetailRequest.java
@@ -1,4 +1,4 @@
-package com.zero.triptalk.planner.dto;
+package com.zero.triptalk.planner.dto.request;
 
 import com.zero.triptalk.place.entity.PlaceRequest;
 import lombok.AllArgsConstructor;
@@ -8,18 +8,15 @@ import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @Builder
-@NoArgsConstructor
 @AllArgsConstructor
-public class UpdatePlannerDetailListRequest {
-    private Long plannerDetailId;
+@NoArgsConstructor
+public class PlannerDetailRequest {
+    private Long id;
     @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm")
     private LocalDateTime date;
     private String description;
     private PlaceRequest placeInfo;
-    private List<String> images;
-
 }
