+++ src/main/java/com/zero/triptalk/planner/dto/request/UpdatePlannerDetailListRequest.java
@@ -1,4 +1,4 @@
-package com.zero.triptalk.planner.dto;
+package com.zero.triptalk.planner.dto.request;
 
 import com.zero.triptalk.place.entity.PlaceRequest;
 import lombok.AllArgsConstructor;
@@ -8,15 +8,18 @@ import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Getter
 @Builder
-@AllArgsConstructor
 @NoArgsConstructor
-public class PlannerDetailRequest {
-    private Long id;
+@AllArgsConstructor
+public class UpdatePlannerDetailListRequest {
+    private Long plannerDetailId;
     @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm")
     private LocalDateTime date;
     private String description;
     private PlaceRequest placeInfo;
+    private List<String> images;
+
 }
