+++ src/main/java/com/zero/triptalk/planner/dto/response/PlannerListResult.java
@@ -1,5 +1,6 @@
-package com.zero.triptalk.planner.dto;
+package com.zero.triptalk.planner.dto.response;
 
+import com.zero.triptalk.planner.dto.response.PlannerListResponse;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
