+++ src/main/java/com/zero/triptalk/planner/repository/CustomPlannerRepository.java
@@ -5,8 +5,8 @@ import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.zero.triptalk.like.entity.QPlannerLike;
-import com.zero.triptalk.planner.dto.PlannerListResponse;
-import com.zero.triptalk.planner.dto.PlannerListResult;
+import com.zero.triptalk.planner.dto.response.PlannerListResponse;
+import com.zero.triptalk.planner.dto.response.PlannerListResult;
 import com.zero.triptalk.planner.entity.QPlanner;
 import com.zero.triptalk.planner.type.SortType;
 import org.springframework.data.domain.Pageable;
