+++ src/main/java/com/zero/triptalk/planner/service/PlannerDetailService.java
@@ -3,8 +3,8 @@ package com.zero.triptalk.planner.service;
 import com.zero.triptalk.exception.custom.PlannerDetailException;
 import com.zero.triptalk.exception.custom.UserException;
 import com.zero.triptalk.image.service.ImageService;
-import com.zero.triptalk.planner.dto.PlannerDetailListResponse;
-import com.zero.triptalk.planner.dto.PlannerDetailResponse;
+import com.zero.triptalk.planner.dto.response.PlannerDetailListResponse;
+import com.zero.triptalk.planner.dto.response.PlannerDetailResponse;
 import com.zero.triptalk.planner.entity.PlannerDetail;
 import com.zero.triptalk.planner.entity.PlannerDetailDocument;
 import com.zero.triptalk.planner.repository.CustomPlannerDetailRepository;
