+++ src/main/java/com/zero/triptalk/planner/service/PlannerService.java
@@ -3,8 +3,8 @@ package com.zero.triptalk.planner.service;
 import com.zero.triptalk.component.RedisUtil;
 import com.zero.triptalk.exception.code.PlannerErrorCode;
 import com.zero.triptalk.exception.custom.PlannerException;
-import com.zero.triptalk.planner.dto.PlannerListResult;
-import com.zero.triptalk.planner.dto.PlannerRequest;
+import com.zero.triptalk.planner.dto.response.PlannerListResult;
+import com.zero.triptalk.planner.dto.request.PlannerRequest;
 import com.zero.triptalk.planner.entity.Planner;
 import com.zero.triptalk.planner.entity.PlannerDocument;
 import com.zero.triptalk.planner.repository.CustomPlannerRepository;
