+++ src/main/java/com/zero/triptalk/search/service/SearchService.java
@@ -4,8 +4,8 @@ import com.zero.triptalk.exception.code.SearchErrorCode;
 import com.zero.triptalk.exception.code.UserErrorCode;
 import com.zero.triptalk.exception.custom.SearchException;
 import com.zero.triptalk.exception.custom.UserException;
-import com.zero.triptalk.planner.dto.PlannerDetailSearchResponse;
-import com.zero.triptalk.planner.dto.PlannerSearchResponse;
+import com.zero.triptalk.planner.dto.response.PlannerDetailSearchResponse;
+import com.zero.triptalk.planner.dto.response.PlannerSearchResponse;
 import com.zero.triptalk.planner.entity.PlannerDetailDocument;
 import com.zero.triptalk.planner.entity.PlannerDocument;
 import com.zero.triptalk.planner.repository.CustomPlannerDetailSearchRepository;
