+++ src/test/java/com/zero/triptalk/application/PlannerApplicationTest.java
@@ -6,9 +6,9 @@ import com.zero.triptalk.image.service.ImageService;
 import com.zero.triptalk.place.entity.Place;
 import com.zero.triptalk.place.entity.PlaceRequest;
 import com.zero.triptalk.place.service.PlaceService;
-import com.zero.triptalk.planner.dto.PlannerDetailListRequest;
-import com.zero.triptalk.planner.dto.PlannerDetailRequest;
-import com.zero.triptalk.planner.dto.PlannerRequest;
+import com.zero.triptalk.planner.dto.request.PlannerDetailListRequest;
+import com.zero.triptalk.planner.dto.request.PlannerDetailRequest;
+import com.zero.triptalk.planner.dto.request.PlannerRequest;
 import com.zero.triptalk.planner.entity.Planner;
 import com.zero.triptalk.planner.entity.PlannerDetail;
 import com.zero.triptalk.planner.service.PlannerDetailService;
