+++ src/test/java/com/zero/triptalk/planner/service/PlannerDetailServiceTest.java
@@ -4,8 +4,8 @@ import com.zero.triptalk.exception.custom.PlannerDetailException;
 import com.zero.triptalk.place.entity.Place;
 import com.zero.triptalk.image.service.ImageService;
 import com.zero.triptalk.place.service.PlaceService;
-import com.zero.triptalk.planner.dto.PlannerDetailRequest;
-import com.zero.triptalk.planner.dto.PlannerDetailResponse;
+import com.zero.triptalk.planner.dto.request.PlannerDetailRequest;
+import com.zero.triptalk.planner.dto.response.PlannerDetailResponse;
 import com.zero.triptalk.planner.entity.Planner;
 import com.zero.triptalk.planner.entity.PlannerDetail;
 import com.zero.triptalk.planner.repository.PlannerDetailRepository;
